testing/web-platform/tests/web-animations/interfaces/KeyframeEffect/processing-a-keyframes-argument-002.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Tests for processing a keyframes argument (easing)</title>
<link rel="help" href="https://w3c.github.io/web-animations/#processing-a-keyframes-argument">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<script src="../../resources/keyframe-utils.js"></script>
<script src="../../resources/easing-tests.js"></script>
<body>
<div id="log"></div>
<div id="target"></div>
<script>
'use strict';

test(function(t) {
  gEasingValueTests.forEach(function(subtest) {
    var easing = subtest[0];
    var expected = subtest[1];
    var effect = new KeyframeEffectReadOnly(target, {
      left: ["10px", "20px"],
      easing: easing
    });
    assert_equals(effect.getKeyframes()[0].easing, expected,
                  "resulting easing for '" + easing + "'");
  });
}, "easing values are parsed correctly when passed to the " +
   "KeyframeEffectReadOnly constructor in a property-indexed keyframe");

test(function(t) {
  gEasingValueTests.forEach(function(subtest) {
    var easing = subtest[0];
    var expected = subtest[1];
    var effect = new KeyframeEffectReadOnly(target, [
      { offset: 0, left: "10px", easing: easing },
      { offset: 1, left: "20px" }
    ]);
    assert_equals(effect.getKeyframes()[0].easing, expected,
                  "resulting easing for '" + easing + "'");
  });
}, "easing values are parsed correctly when passed to the " +
   "KeyframeEffectReadOnly constructor in regular keyframes");

test(function(t) {
  gInvalidEasings.forEach(invalidEasing => {
    assert_throws(new TypeError, () => {
      new KeyframeEffectReadOnly(target, { easing: invalidEasing });
    }, `TypeError is thrown for easing '${invalidEasing}'`);
  });
}, 'invalid easing values are correctly rejected when passed to the ' +
   'KeyframeEffectReadOnly constructor in regular keyframes');

test(function(t) {
  gInvalidEasings.forEach(invalidEasing => {
    assert_throws(new TypeError, () => {
      new KeyframeEffectReadOnly(target, [{ easing: invalidEasing }]);
    }, `TypeError is thrown for easing '${invalidEasing}'`);
  });
}, 'invalid easing values are correctly rejected when passed to the ' +
   'KeyframeEffectReadOnly constructor in a keyframe sequence');

</script>
